new_installer: revert DatabaseAction and ChildInfo
A prior change speculatively renamed ChildInfo -> PosixChildInfo and moved it
to new_installer_posix.py (and DatabaseAction to new_installer_base.py),
anticipating per-platform subclasses that did not materialize.
